requirements.txt
Flask==2.3.5
Flask-Login==0.7.0
Flask-WTF==1.1.1
SQLAlchemy==2.1.2
WTForms==3.0.1
python-dotenv==1.0.0
Flask-Mail==0.9.1
pandas==2.2.3
PyJWT==2.8.0
pytest==7.4.0
pytest-flask==1.3.0
python-multipart==0.0.5
